New test cases, some failing test cases were fixed due to logical errors.
Changes made:

id in database is of type long with the sqlite library, not int. changing this everywhere. functionality across the app was tested to ensure no regressions were caused
by this *major* change.

pls check:
https://stackoverflow.com/questions/7575166/android-sqlite-get-last-insert-row-id

there was a logical error when trying to remove punches for a user, when a user is deleted. re-wrote the deletion method (a test-case was failing)
more unit test cases added